fix: stabilize benchmark runs with registration, loading, and config fixes
- Use force=True on all register_module() calls for re-registration safety
- Rewrite CLIP backbone weight loading to handle TorchScript in PyTorch>=2.6
- Fix GLASS detector legacy train_step infinite recursion
- Add VISA auto-detection in benchmark.py
- Fix aaclip VISA config (was pointing to MVTec data)
- Remove compose_ad configs (in skip list)
- Add new benchmark/training tools (speed, ast, regad, vitad)
- Update .gitignore for groundingdino/, agent docs

+++ baoiad/evaluation/aupro.py
@@ -90,7 +90,7 @@ def compute_aupro(gt_masks: np.ndarray | Sequence[np.ndarray], pred_maps: np.nda
     cca = _make_global_region_labels(np.asarray(cca, dtype=object))
 
     # Step 2: Flatten
-    labels = np.concatenate([np.asarray(item).reshape(-1) for item in cca])
+    labels = np.concatenate([np.asarray(item, dtype=np.int64).reshape(-1) for item in cca])
     preds_flat = np.concatenate([pred_map.reshape(-1).astype(np.float64) for pred_map in pred_list])
 
     # Step 3: Compute contributions
